requirements.txt
annotated-types==0.7.0
anyio==4.10.0
black==25.1.0
certifi==2025.8.3
click==8.2.1
fastapi==0.116.1
flake8==7.3.0
h11==0.16.0
httptools==0.6.4
idna==3.10
influxdb-client==1.49.0
iniconfig==2.1.0
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
reactivex==4.0.4
redis==6.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# MCP (Model Context Protocol) Dependencies for Commercial Collectors
jsonrpc-requests==0.4.0
requests-toolbelt==1.0.0
aiohttp==3.9.1
